// src/users/users.controller.interface.ts
import { NextFunction, Request, Response } from 'express';

export interface IUsersController {
	login: (req: Request, res: Response, next: NextFunction) => void;
	register: (req: Request, res: Response, next: NextFunction) => void;
	info: (req: Request, res: Response, next: NextFunction) => void;
	//getUserInfo: (req: Request, res: Response, next: NextFunction) => void;
}